Hack to sanitize non-utf8 strings
This commit simply replaces any non-utf8 characters in the ticket
description with the unicode replacement character.

This probably isn't the best way to handle this; the strings coming
from redmine are probably encoded in a known character set (probably
latin-1), and if so, we could just do a "force_encoding" to that
character set.  However, this will at least ensure that any weird
characters that appear in a ticket description won't prevent that
ticket from being copied over.