Data sources have control over existance checks
Previously the data synchronization assumed that the data would only
ever live in the one list that the cards are created in. This is not
always the case and now the data sources can create the synchronization
data with the strategy of how to discover if the data exists already.
This allows github data to be moved into other columns with matches the
possible workflow for it.