Add Github OAuth token support to avoid rate limits
Without this patch applied the implementation does not use any
authentication when communicating with the Github API.  This is a
problem because Github has very low rate limits (100/hour) for
unauthenticated requests.

This patch addresses the problem by accepting a Github OAuth token as a
configuration parameter.  Using this token increases the rate limit to
5000 requests per hour.

A token may be obtained like so:

    GITHUB_USER=gepetto-bot
    curl -u $GITHUB_USER -d '
    {
      "scopes":["repo"],
      "note":"redmine-trello",
      "note_url":"https://github.com/cprice-puppet/redmine-trello"
    }' https://api.github.com/authorizations

A token may be revoked at https://github.com/settings/applications